Start moving interaction code into regular submodule
As preparation for putting test interface classes alongside the st
element function implementations.
